Add files via upload
This Dashboard is a road accident analysis  where we Identified primay KPIs such as Total CY Casualities and rate of decrease of current year and previous year, CY accidents, CY Fatal, Serious and slight casual. Secondary KPIS such as the Casaul vehicle type was identified.

This was used in order to get more insights about the road type, Location of these accidents (Rural and Urban) based on the different coordinate points and finally we could observe this on a CY casuality and PY causuality graph. 

We concluded by providing filtreing functionalities to be able to sort by weather conditions and road surface conditions.